:truck: Moving NodesWindow to another module
