+++ dial_gui/widgets/editor_tabwidget/editor_tabwidget.py
@@ -3,18 +3,19 @@
 from typing import TYPE_CHECKING
 
 import dependency_injector.providers as providers
-from dial_gui.node_editor import NodeEditorWindowFactory
-from dial_gui.project import ProjectManagerGUISingleton
-from dial_gui.widgets.node_panels import NodesWindow
 from PySide2.QtCore import QEvent
 from PySide2.QtGui import QIcon, QPixmap
 from PySide2.QtWidgets import QStackedWidget, QTabBar, QTabWidget, QWidget
+
+from dial_gui.node_editor import NodeEditorWindowFactory
+from dial_gui.node_editor.nodes_windows import NodesWindow
+from dial_gui.project import ProjectManagerGUISingleton
 
 from .tab_settings_dialog import TabSettingsDialog
 
 if TYPE_CHECKING:
     from dial_gui.project import ProjectManagerGUI
-    from dial_gui.widgets.node_panels import NodesWindowsManager
+    from dial_gui.node_editor.nodes_windows import NodesWindowsManager
 
 
 class CustomTabWidget(QTabWidget):
