--- dial_gui/widgets/node_panels/__init__.py
@@ -1,5 +1,0 @@
-# vim: ft=python fileencoding=utf-8 sts=4 sw=4 et:
-
-from .node_panels import NodesWindow, NodesWindowsManager, NodesWindowsManagerFactory
-
-__all__ = ["NodesWindow", "NodesWindowsManager", "NodesWindowsManagerFactory"]
